+++ spark/src/main/scala/ai/chronon/spark/databricks/DatabricksTableUtils.scala
@@ -1,6 +1,7 @@
-package ai.chronon.spark
+package ai.chronon.spark.databricks
 
 import ai.chronon.api.Constants
+import ai.chronon.spark.BaseTableUtils
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.{Column, SparkSession}
